requirements.txt
fastapi
smart-api-bridge